+++ packages/gatsby-theme-podcast/gatsby-node.js
@@ -1,8 +1,5 @@
 const fs = require("fs");
 const path = require("path");
-const crypto = require("crypto");
-const Parser = require("rss-parser");
-const parser = new Parser();
 
 exports.onPreBootstrap = ({ reporter }) => {
   const dirs = [
@@ -19,92 +16,6 @@ exports.onPreBootstrap = ({ reporter }) => {
     }
   });
 };
-
-const createContentDigest = obj =>
-  crypto
-    .createHash("md5")
-    .update(JSON.stringify(obj))
-    .digest("hex");
-
-const generateRSSItemSlug = title => {
-  return (
-    title
-      // remove non-alphanumeric characters, except spaces
-      .replace(/[^\w\s]/gi, "")
-      .toLowerCase()
-      .split(" ")
-      .join("-")
-  );
-};
-
-const durationToMinutes = durationInSeconds => {
-  return Math.floor(durationInSeconds / 60);
-};
-
-// hackity hack. this extracts the first paragraph from the description
-const extractFirstParagraph = html => {
-  return html.match(/<p>(.*?)<\/p>/)[0];
-};
-
-async function sourceNodes({ actions }, { rssSource = "" }) {
-  if (!rssSource) {
-    console.log(`gatsby-theme-podcast requires an RSS feed`);
-    return;
-  }
-  const { createNode } = actions;
-  const data = await parser.parseURL(rssSource);
-
-  if (!data) {
-    return;
-  }
-
-  // Create nodes for top-level RSS feed info
-  const { title, description, link, image, items } = data;
-
-  // Prepare RSS item nodes
-  const preparedRssItemNodes = items.map(rssItem => {
-    let slug = generateRSSItemSlug(rssItem.title);
-    let excerpt = extractFirstParagraph(rssItem.content);
-    let duration = durationToMinutes(rssItem.itunes.duration);
-    const node = Object.assign({}, rssItem, {
-      id: rssItem.link,
-      title: rssItem.title,
-      content: rssItem.content,
-      link: rssItem.link,
-      excerpt,
-      duration,
-      slug,
-      parent: null
-    });
-    node.internal = {
-      type: "rssFeedItem",
-      contentDigest: createContentDigest(node)
-    };
-    return node;
-  });
-
-  const feedInfo = {
-    id: link,
-    title,
-    description,
-    link,
-    imageUrl: image.url,
-    items___NODE: preparedRssItemNodes.map(node => node.id),
-    parent: null
-  };
-
-  feedInfo.internal = {
-    type: "rssFeedInfo",
-    contentDigest: createContentDigest(feedInfo)
-  };
-
-  createNode(feedInfo);
-
-  // Actually create nodes for RSS items
-  preparedRssItemNodes.forEach(rssItemNode => createNode(rssItemNode));
-}
-
-exports.sourceNodes = sourceNodes;
 
 exports.createPages = async function createPages({ graphql, actions }) {
   const { createPage } = actions;
@@ -146,49 +57,21 @@ exports.createPages = async function createPages({ graphql, actions }) {
   });
 };
 
-// *
-// * @TODO this is all now outdated
-// * These files should only exist in the site/theme starter
-// * Need to be pulled in (if exists) to corresponding episode
-// *
-let userCreatedOwnEpisodes = false;
+// Connect mdx notes/transcripts to episode
+exports.onCreateNode = async ({ node, actions, getNodesByType }) => {
+  const { createParentChildLink } = actions;
 
-exports.onCreateNode = ({ node, actions, getNode }) => {
-  const { createNodeField } = actions;
+  if (node.internal.type !== `Mdx`) {
+    return;
+  }
 
-  if (node.internal.type === `Mdx`) {
-    // create source field
-    const fileNode = getNode(node.parent);
+  const rssItemNodes = getNodesByType(`rssFeedItem`);
 
-    const source = fileNode.sourceInstanceName;
+  const filtered = rssItemNodes.find(rssItemNode => {
+    return rssItemNode.link === node.frontmatter.boop;
+  });
 
-    createNodeField({
-      node,
-      name: `source`,
-      value: source
-    });
-
-    const eligibleEpisodeSources = [
-      "podcast-demo-episodes",
-      "podcast-episodes"
-    ];
-
-    if (eligibleEpisodeSources.includes(source)) {
-      if (source === "podcast-episodes") {
-        userCreatedOwnEpisodes = true;
-      }
-
-      if (userCreatedOwnEpisodes && source === "podcast-demo-episodes") {
-        return;
-      }
-
-      // create slug for episode pages
-      const value = path.parse(node.fileAbsolutePath).name;
-      createNodeField({
-        name: `slug`,
-        node,
-        value
-      });
-    }
+  if (filtered) {
+    createParentChildLink({ parent: filtered, child: node });
   }
 };
