+++ packages/gatsby-theme-podcast/package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
+    "@amber1ey/gatsby-source-rss-feed": "*",
     "@mdx-js/mdx": "^1.0.4",
     "@mdx-js/react": "^1.0.2",
     "gatsby-image": "^2.0.34",
