+++ packages/wip-gatsby-source-rss-feed/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@amber1ey/gatsby-source-rss-feed",
+  "private": true,
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "rss-parser": "^3.7.2"
+  }
+}
