+++ podcast-site/content/episodes/0-intro.mdx
@@ -1,7 +1,7 @@
 ---
-title: Woof
+title: Meow
 boop: "https://anchor.fm/fullstack-health/episodes/0-Intro-e4egsi"
-ygatsby: eh
+ygatsby: meh
 ---
 
 ## Links
